requirements.txt
Pillow==9.0.1
Flask==2.1.1
python-dotenv==0.20.0